requirements.txt
Django==1.9.1
django-filter==0.11.0
djangorestframework==3.3.1
Markdown==2.6.5
mysqlclient==1.3.7
numpy==1.10.1
wheel==0.24.0
